-feat: added bookmarks and currently readings tabs to userlayout

+++ src/miranaSolution.Services/Core/Books/BookService.cs
@@ -14,7 +14,7 @@ public class BookService : IBookService
     private readonly MiranaDbContext _context;
     private readonly IImageSaver _imageSaver;
     private readonly IValidatorProvider _validatorProvider;
-    
+
     public BookService(MiranaDbContext context, IImageSaver imageSaver, IValidatorProvider validatorProvider)
     {
         _context = context;
@@ -25,53 +25,42 @@ public class BookService : IBookService
     public async Task<GetBookByIdResponse> GetBookByIdAsync(GetBookByIdRequest request)
     {
         var book = await _context.Books.FindAsync(request.BookId);
-        if (book is null)
-        {
-            return new GetBookByIdResponse(null);
-        }
+        if (book is null) return new GetBookByIdResponse(null);
 
         var bookVm = MapBookIntoBookVm(book);
 
         var response = new GetBookByIdResponse(bookVm);
-        
+
         return response;
     }
 
     public async Task<GetBookBySlugResponse> GetBookBySlugAsync(GetBookBySlugRequest request)
     {
         var book = await _context.Books.FirstOrDefaultAsync(x => x.Slug.Equals(request.Slug));
-        if (book is null)
-        {
-            return new GetBookBySlugResponse(null);
-        }
+        if (book is null) return new GetBookBySlugResponse(null);
 
         var bookVm = MapBookIntoBookVm(book);
 
         var response = new GetBookBySlugResponse(bookVm);
-        
+
         return response;
     }
-    
+
     /// <exception cref="UserNotFoundException">
-    /// Thrown when the user with given Id does not exist
+    ///     Thrown when the user with given Id does not exist
     /// </exception>
     /// <exception cref="BookAlreadyExistsException">
-    /// Thrown when the book with given Slug already exists
+    ///     Thrown when the book with given Slug already exists
     /// </exception>
     public async Task<CreateBookResponse> CreateBookAsync(CreateBookRequest request)
     {
         _validatorProvider.Validate(request);
 
         if (!await _context.Authors.AnyAsync(x => x.Id == request.AuthorId))
-        {
             throw new AuthorNotFoundException("The author with given Id does not exist.");
-        }
 
         var book = await _context.Books.FirstOrDefaultAsync(x => x.Slug.Equals(request.Slug));
-        if (book is not null)
-        {
-            throw new BookAlreadyExistsException("The book with given Slug already exists.");
-        }
+        if (book is not null) throw new BookAlreadyExistsException("The book with given Slug already exists.");
 
         book = new Book
         {
@@ -81,7 +70,7 @@ public class BookService : IBookService
             IsRecommended = request.IsRecommended,
             Slug = request.Slug,
             IsDone = request.IsDone,
-            ThumbnailImage = 
+            ThumbnailImage =
                 await _imageSaver.SaveImageAsync(request.ThumbnailImage, request.ThumbnailImageExtension),
             AuthorId = request.AuthorId
         };
@@ -90,32 +79,27 @@ public class BookService : IBookService
         await _context.SaveChangesAsync();
 
         var bookVm = MapBookIntoBookVm(book);
-        
+
         var response = new CreateBookResponse(bookVm);
 
         return response;
     }
-    
+
     /// <exception cref="BookNotFoundException">
-    /// Thrown when the book with given Id is not found
+    ///     Thrown when the book with given Id is not found
     /// </exception>
     public async Task<UpdateBookResponse> UpdateBookAsync(UpdateBookRequest request)
     {
         _validatorProvider.Validate(request);
-        
+
         var book = await _context.Books.FindAsync(request.BookId);
-        if (book is null)
-        {
-            throw new BookNotFoundException("The book with given Id does not exists.");
-        }
-        
+        if (book is null) throw new BookNotFoundException("The book with given Id does not exists.");
+
         // Validate the new slug must be unique
         if (await _context.Books.AnyAsync(
                 x => !x.Slug.Equals(book.Slug) && x.Slug.Equals(request.Slug)))
-        {
             throw new BookAlreadyExistsException("The book with given Slug already exists.");
-        }
-        
+
         book.Name = request.Name;
         book.ShortDescription = request.ShortDescription;
         book.LongDescription = request.LongDescription;
@@ -132,27 +116,22 @@ public class BookService : IBookService
         }
 
         foreach (var item in request.GenreCheckboxItems)
-        {
             if (!item.IsChecked)
             {
                 var bookGenre = await _context.BookGenres
                     .FirstOrDefaultAsync(x => x.BookId == book.Id && x.GenreId == item.Id);
-                if (bookGenre is not null)
-                {
-                    _context.BookGenres.Remove(bookGenre);
-                }
+                if (bookGenre is not null) _context.BookGenres.Remove(bookGenre);
             }
             else
             {
-                var bookGenre = new BookGenre()
+                var bookGenre = new BookGenre
                 {
                     BookId = book.Id,
                     GenreId = item.Id
                 };
-        
+
                 await _context.BookGenres.AddAsync(bookGenre);
             }
-        }
 
         await _context.SaveChangesAsync();
 
@@ -163,18 +142,15 @@ public class BookService : IBookService
     }
 
     /// <exception cref="BookNotFoundException">
-    /// Thrown when the book with given Id is not found
+    ///     Thrown when the book with given Id is not found
     /// </exception>
     public async Task DeleteBookAsync(DeleteBookRequest request)
     {
         var book = await _context.Books.FindAsync(request.BookId);
-        if (book is null)
-        {
-            throw new BookNotFoundException("The book with given Id does not exist.");
-        }
+        if (book is null) throw new BookNotFoundException("The book with given Id does not exist.");
 
         await _imageSaver.DeleteImageIfExistAsync(book.ThumbnailImage);
-        
+
         _context.Books.Remove(book);
         await _context.SaveChangesAsync();
     }
@@ -186,51 +162,49 @@ public class BookService : IBookService
             .ToListAsync();
 
         var bookVms = books.Select(MapBookIntoBookVm).ToList();
-        
+
         return new GetRecommendedBooksResponse(bookVms);
     }
 
     public async Task<GetAllBooksResponse> GetAllBooksAsync(GetAllBooksRequest request)
     {
         var query = _context.Books.AsQueryable();
-        
+
         // Apply books filter by keyword
         if (!string.IsNullOrEmpty(request.Keyword))
-        {
             query = query.Where(x =>
                 x.Name.Contains(request.Keyword)
                 || x.ShortDescription.Contains(request.Keyword)
                 || x.LongDescription.Contains(request.Keyword));
-        }
 
         // Apply books filter by genre
         if (!string.IsNullOrEmpty(request.GenreIds))
         {
             var genreIds = request.GenreIds.Split(",").Select(int.Parse);
             foreach (var genreId in genreIds)
-            {
                 query = query
                     .Include(x => x.BookGenres)
                     .Where(x => x.BookGenres.Select(_ => _.GenreId).Contains(genreId));
-            }
         }
-        
-        // Apply books filter by status
-        if (request.IsDone.HasValue)
+
+        if (request.AuthorId.HasValue)
         {
-            query = query.Where(x => x.IsDone == request.IsDone);
+            query = query.Where(x => x.AuthorId == request.AuthorId);
         }
-        
+
+        // Apply books filter by status
+        if (request.IsDone.HasValue) query = query.Where(x => x.IsDone == request.IsDone);
+
         var pageSize = request.PagerRequest.PageSize;
         var pageIndex = request.PagerRequest.PageIndex;
-        
+
         var totalBooks = await query.CountAsync();
-        
+
         var books = await query
             .Skip((pageIndex - 1) * pageSize)
             .Take(pageSize)
             .ToListAsync();
-        
+
         var bookVms = books.Select(MapBookIntoBookVm).ToList();
 
         var pagerResponse = new PagerResponse(
@@ -241,7 +215,7 @@ public class BookService : IBookService
         var response = new GetAllBooksResponse(
             bookVms,
             pagerResponse);
-        
+
         return response;
     }
 
@@ -270,7 +244,7 @@ public class BookService : IBookService
 
         var authorName = book.Author!.Name;
         var genres = book.BookGenres
-            .Select(bookGenre => 
+            .Select(bookGenre =>
             {
                 _context.Attach(bookGenre);
                 _context.Entry(bookGenre)
@@ -293,7 +267,8 @@ public class BookService : IBookService
             authorName,
             genres,
             book.IsDone,
-            book.AuthorId);
+            book.AuthorId,
+            book.ViewCount);
 
         return bookVm;
     }
